Port last legacy widgets and remove _legacy folder
