+++ src/spikeinterface/widgets/__init__.py
@@ -3,9 +3,3 @@ from .widget_list import *
 # general functions
 from .utils import get_some_colors, get_unit_colors, array_to_image
 from .base import set_default_plotter_backend, get_default_plotter_backend
-
-
-# we keep this to keep compatibility so we have all previous widgets
-# except the one that have been ported that are imported
-# with "from .widget_list import *" in the first line
-from ._legacy_mpl_widgets import *
