--- src/spikeinterface/widgets/_legacy_mpl_widgets/__init__.py
@@ -1,8 +1,0 @@
-from .multicompgraph import (
-    plot_multicomp_graph,
-    MultiCompGraphWidget,
-    plot_multicomp_agreement,
-    MultiCompGlobalAgreementWidget,
-    plot_multicomp_agreement_by_sorter,
-    MultiCompAgreementBySorterWidget,
-)
